refactor hero example block
